Update typing for `incr`/`decr`
Allow the count for these functions to be `float`. This was inspired by
an existing test that used such a value, I also tested via a run of

    $ docker run -d \
     --name graphite \
     --restart=always \
     -p 80:80 \
     -p 2003-2004:2003-2004 \
     -p 2023-2024:2023-2024 \
     -p 8125:8125/udp \
     -p 8126:8126 \
     graphiteapp/graphite-statsd
   $ echo -n "example:5.5|c" | nc -w 1 -u 127.0.0.1 8125;

The dashboard at http://localhost/dashboard then happily displayed the
value 5.5 (I'm assuming there's not some incompatibility with `statsd`
in the pipeline)

+++ statsd/client/base.py
@@ -40,11 +40,11 @@ class StatsClientBase:
             delta = delta.total_seconds() * 1000.0
         self._send_stat(stat, "%0.6f|ms" % delta, rate)
 
-    def incr(self, stat: str, count: int = 1, rate: float = 1) -> None:
+    def incr(self, stat: str, count: float = 1, rate: float = 1) -> None:
         """Increment a stat by `count`."""
         self._send_stat(stat, "%s|c" % count, rate)
 
-    def decr(self, stat: str, count: int = 1, rate: float = 1) -> None:
+    def decr(self, stat: str, count: float = 1, rate: float = 1) -> None:
         """Decrement a stat by `count`."""
         self.incr(stat, -count, rate)
